Complete implementation of comments on pictures
The implementaiton is complete. However, the comment controller
test for creating a picture is broken and has been disabled.

The reason I've committed here is that I've realised there's a
possible refactor I can do that will massively simplify the
writing of the test that I'm struggling with, so I want to explore
that separately.