Write comments controller create test so it actually works
I had a huge amount of difficulty writing this test, and it turns
out I had got confused about the create method called on a model
and the create method called on a controller when a post route
is activated.

Once I'd sorted this out in my head, the test didn't take long to
write. Unfortunately it meant that I didn't follow TDD, with the
code coming before the test, but I'm confident that this won't
happen anymore based on what I've learned from going through the
process this time.